+++ DNN/NeuralNet/SGD.cs
@@ -1,11 +1,12 @@
+using System;
 namespace NeuralNet
 {
     public class SGD : Optimizer
     {
-        public SGD(int lr) : base(lr){}
+        public SGD(double lr) : base(lr){}
 
         public override void Step(Module module){
-            foreach(Parameter param in module.Parameters()){
+            foreach(Parameter param in module.Parameters()){            
                 param.Data -= param.Grad.Data * LearningRate;
             }
         }
